requirements.txt
# Core web framework and dashboard
streamlit>=1.28.0

# Data analysis and visualization
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0

# HTTP requests and networking
requests>=2.31.0
aiohttp>=3.8.0

# Security tools integration
python-nmap>=0.7.1
shodan>=1.29.1
dnspython>=2.4.0

# Report generation
reportlab>=4.0.0
jinja2>=3.1.0
weasyprint>=59.0

# Data validation and parsing
pydantic>=2.0.0
beautifulsoup4>=4.12.0

# Configuration and environment
python-dotenv>=1.0.0

# AI/LLM integration
openai>=1.0.0
anthropic>=0.3.0

# Local LLM integration (Llama 3.1 8B)
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99

# System utilities
psutil>=5.9.0

# Optional: Enhanced functionality
# uncomment if needed
# scapy>=2.5.0
# paramiko>=3.0.0
# ftplib
# smtplib
